added branch, condensed some files along with making header one single file with extensions (more modularity)
