+++ src/components/frontend/UI/headers/header.jsx
@@ -1,16 +1,16 @@
 import React, { useState, useEffect } from 'react';
-import '../App.css';
-import { Darklight } from './darklight';
+
+import { Darklight } from '../../darklight';
 import { useCookies } from 'react-cookie';
 
-function AboutHeader() {
+function Header( {description, image} ) {
   const [cookies, setCookie] = useCookies(['darkMode']);
   const [darkMode, setDarkMode] = useState(false);
 
   useEffect(() => {
     const savedDarkMode = cookies.darkMode === 'true';
     setDarkMode(savedDarkMode);
-  }, [cookies.darkMode]);
+  }, [cookies.darkMode, description]);
 
   const toggleDarkMode = () => {
     const newMode = !darkMode;
@@ -18,12 +18,14 @@ function AboutHeader() {
     setCookie('darkMode', newMode.toString(), { path: '/' });
   };
 
+  <Darklight darkMode={darkMode} toggleDarkMode={toggleDarkMode} showToggle={false} />
+
   const styles = () => (
     <style jsx>{`
-      .about-header-container,
-      .about-header-container-dark {
-        background: url('/images/about-me.jpg') no-repeat;
-        height: 101vh;
+      .header-container,
+      .header-container-dark {
+        background: url('${image}') no-repeat;
+        height: 100vh;
         width: 100%;
         display: flex;
         flex-direction: column;
@@ -35,74 +37,94 @@ function AboutHeader() {
         user-select: none;
       }
       
-      .about-header-container > h1,
-      .about-header-container-dark > h1 {
+      .header-container > h1,
+      .header-container-dark > h1 {
         font-size: 100px;
         margin-top: 250px;
       }
       
-      .about-header-container > h1 {
+      .header-container > h1 {
         color: #9AC2E6;
       }
       
-      .about-header-container-dark > h1 {
+      .header-container-dark > h1 {
         color: #B39DDB;
         font-size: 100px;
         margin-top: 250px;
       }
       
-      .about-header-container > h1 span {
+      .header-container > h1 span {
         position: relative;
         display: inline-block;
       }
       
       @media screen and (min-width: 961px) {
-        .about-header-container,
-        .about-header-container-dark {
+        .header-container,
+        .header-container-dark {
           object-fit: scale-down;
         }
       }
       
       @media screen and (max-width: 960px) {
-        .about-header-container > h1,
-        .about-header-container-dark > h1 {
+        .header-container > h1,
+        .header-container-dark > h1 {
           font-size: 70px;
           margin-top: -150px;
         }
-        .about-header-container,
-        .about-header-container-dark {
+        .header-container,
+        .header-container-dark {
           background-size: cover;
         }
       }
       
       @media screen and (max-width: 768px) {
-        .about-header-container > h1,
-        .about-header-container-dark > h1 {
-          font-size: 50px;
+        .header-container > h1,
+        .header-container-dark > h1 {
+          font-size: auto;
           margin-top: -100px;
         }
       }
       
       @media screen and (max-width: 500px) {
-        .about-header-container,
-        .about-header-container-dark {
+        .header-container,
+        .header-container-dark {
           background-size: cover;
         }
       }
+
+      @media screen and (max-height: 600px) {
+        .header-container > h1,
+        .header-container-dark > h1 {
+          font-size: 50px;
+          margin-top: -100px;
+        }
+      }
     `}</style>
+    );
+
+    /*
+  return (
+    <>
+      <div className={darkMode ? 'header-container-dark' : 'header-container'}>
+        <h1>
+          <span>software</span>
+          <hr />
+          <span>engineer</span>
+        </h1>
+      </div>
+      {styles()}
+    </>
   );
+  */
 
   return (
     <>
-      <div className={`about-header-container${darkMode ? '-dark' : ''}`}>
-        <h1>
-          <span>about me</span>
-        </h1>
-        <Darklight darkMode={darkMode} toggleDarkMode={toggleDarkMode} showToggle={false} />
-        {styles()}
+      <div className={darkMode ? 'header-container-dark' : 'header-container'}>
+          <h1 dangerouslySetInnerHTML={{ __html: description }} />
       </div>
+      {styles()}
     </>
   );
 }
 
-export default AboutHeader;
+export default Header;
