+++ src/components/frontend/UI/pages/about-me.js
@@ -1,12 +1,10 @@
 import React from 'react';
-import '../../App.css';
-import AboutHeader from '../about-header';
-//import Cards2 from '../about-cards';
-import Footer from '../footer';
-import AboutCarousel from '../about-carousel';
-import PokemonGrid from '../pokemonGrid';
-import IGDB from '../IGDB';
-import Fruit from '../fruit';
+import AboutHeader from '../headers/about-header';
+import Footer from '../../footer';
+import AboutCarousel from '../../about-carousel';
+import PokemonGrid from '../../pokemon/pokemonGrid';
+import IGDB from '../../games/IGDB';
+import Fruit from '../../fruit';
 //import Music from '../music';
 
 function AboutMe() {
