+++ src/components/frontend/UI/pages/home.js
@@ -1,7 +1,7 @@
 import React from 'react';
-import Header from '../header';
-import Cards from '../cards';
-import Footer from '../footer';
+import Header from '../headers/main-header';
+import Cards from '../../cards';
+import Footer from '../../footer';
 //import FloatingButton from '../floatingButton';
 
 function home() {
