+++ src/components/frontend/about-carousel.jsx
